+++ modules/integrations/cats/src/main/scala/tethys/cats/writers/CatsWriters.scala
