+++ modules/integrations/enumeratum/src/main/scala-2/tethys/enumeratum/TethysEnum.scala
