+++ modules/integrations/enumeratum/src/test/scala-2/tethys/enumeratum/Direction.scala
